#9986 - Changed PulsarSink to use ProducerConfig for defaults and got tests to pass in PulsarSinkTest

+++ pulsar-functions/utils/src/main/java/org/apache/pulsar/functions/utils/functions/ClusterFunctionProducerDefaults.java
@@ -43,7 +43,7 @@ public class ClusterFunctionProducerDefaults {
     private HashingScheme hashingScheme;
     private MessageRoutingMode messageRoutingMode;
 
-    private int batchingMaxPublishDelay;
+    private long batchingMaxPublishDelay;
 
     public ClusterFunctionProducerDefaults(boolean batchingDisabled, boolean chunkingEnabled, boolean blockIfQueueFullDisabled,
                                            String compressionType, String hashingScheme, String messageRoutingMode,
